Fix the _imagingft check to have no presumption about import paths.
With Pillow 2.1.0 the _imagingft import location has changed, it is now
within the PIL package.
Since we were directly importing it, we accidentally presumed that it is
missing, but it wasn't.

By reusing the Pillow check and logging a better message we have compatibility
with older and newer Pillow versions.