Fix error when portrait is stored as Pdata object.
Sometimes the portrait is stored as Pdata object instead of as string,
which causes an AttributeError "read" when trying to scale it.

It is unclear why those images are stored differently.
Although it happend with a PNG image, not all PNG images are affected,
nor is it the file size relevant.