Avoid writes on read - create default avatar on login.
To prevent "writes on read", I want to change the behavior of
ftw.avatar. Instead of creating the default avatar on the portrait
access, the avatar gets created after a successfull login.

With a few exceptions, this should change very little for the
user. Only if avatars are displayed by users who have never logged in
before.